+++ src/main/kotlin/icu/windea/pls/lang/util/ParadoxComplexEnumValueManager.kt
@@ -245,6 +245,15 @@ object ParadoxComplexEnumValueManager {
         return ParadoxLocalisationSearch.search(name, selector).find()
     }
 
+    fun getNameLocalisationFromExtendedConfig(name: String, enumName: String, contextElement: PsiElement): ParadoxLocalisationProperty? {
+        val hint = getHintFromExtendedConfig(name, enumName, contextElement) //just use file as contextElement here
+        if (hint.isNullOrEmpty()) return null
+        val hintLocalisation = ParadoxLocalisationElementFactory.createProperty(contextElement.project, "hint", hint)
+        //it's necessary to inject fileInfo here (so that gameType can be got later)
+        hintLocalisation.containingFile.virtualFile.putUserData(PlsKeys.injectedFileInfo, contextElement.fileInfo)
+        return hintLocalisation
+    }
+
     fun getHintFromExtendedConfig(name: String, enumName: String, contextElement: PsiElement): String? {
         if (name.isEmpty()) return null
         val gameType = selectGameType(contextElement) ?: return null