+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxComplexEnumValueLocalizedNameHintsProvider.kt
@@ -81,19 +81,8 @@ class ParadoxComplexEnumValueLocalizedNameHintsProvider : ParadoxScriptHintsProv
     }
 
     private fun getNameLocalisationToUse(name: String, enumName: String, file: PsiFile): ParadoxLocalisationProperty? {
-        run {
-            val hint = ParadoxComplexEnumValueManager.getHintFromExtendedConfig(name, enumName, file) //just use file as contextElement here
-            if (hint.isNullOrEmpty()) return@run
-            val hintLocalisation = ParadoxLocalisationElementFactory.createProperty(file.project, "hint", hint)
-            //it's necessary to inject fileInfo here (so that gameType can be got later)
-            hintLocalisation.containingFile.virtualFile.putUserData(PlsKeys.injectedFileInfo, file.fileInfo)
-            return hintLocalisation
-        }
-        run {
-            val nameLocalisation = ParadoxComplexEnumValueManager.getNameLocalisation(name, file, ParadoxLocaleManager.getPreferredLocaleConfig())
-            if (nameLocalisation == null) return@run
-            return nameLocalisation
-        }
+        ParadoxComplexEnumValueManager.getNameLocalisationFromExtendedConfig(name, enumName, file)?.let { return it }
+        ParadoxComplexEnumValueManager.getNameLocalisation(name, file, ParadoxLocaleManager.getPreferredLocaleConfig())?.let { return it }
         return null
     }
 }